% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignmentB2.R
\name{Data_overview}
\alias{Data_overview}
\title{Group & Summarize}
\usage{
Data_overview(data, x, y)
}
\arguments{
\item{data}{The data that you will be using for this function.}

\item{x}{The variable in which you will be grouping the data by.}

\item{y}{The numerical variable of interest. This variable MUST be a numerical variable for this function to work.}
}
\value{
This function will return descriptive characteristics of the numerical variable (y), where the range (max - min), mean, and the standard deviation will be calculated.
}
\description{
This function will group your data by a specified variable (x), and create three columns:
Max_Minus_Min (The maximum-minimum value of the numerical variable y), Mean of y, and the Standard deviation of y.
Note that this second variable (y) MUST be a numerical variable for this function to work, as the range (max - min), mean, and standard deviation of categorical variables cannot be calculated.
}
\examples{
Data_overview(palmerpenguins::penguins, island, flipper_length_mm)
Data_overview(iris, Species, Sepal.Length)
}
